Add routes for confirmation system
We need two routes to link to the same system, one by Parse ID and one by
email. Email is necessary for quick, on the fly day-of confirmations, as we
won't allow them to check in until they've confirmed (accepting a waiver and
some other stuff). The email route should only be served on the day-of.