Setup submission script for confirmations
Needed to add the JS file for submitting the AJAX request, but also
restructured the POST request to just `/confirm`. This is because I supply the
hackerId through a hidden HTML form field, so there is no need to worry about
posting to a route with an id parameter. Although, this way might be against
REST, so we should look into that later. It's 4am atm though, so fuck it.