Create a method to save a Confirmation
I'm not exactly sure how to override the save() method and still use the
Parse SDK to save the data once that function is overwritten. Instead, to
be safe for now, I simply named the function saveIt() so that there aren't
any issues. We'll need to come back to that later and fix it.