Add a lightweight redirect on invalid confirmation id
Basically, when we render the confirmation page, we have to check whether
or not the hacker supplied their phone # when they registered. Because of
this, we are required to check through Parse if the `phone` field is set,
and this has a brilliant side-effect of sending an error if an invalid id
was supplied (through URI params). Thus, we simply redirect to `/no` in this
case, which is a nice way imo to redirect to the 404 page.