Begin work on figuring out a way to easily superclass standard methods
For example, every model will likely use a "findAll()" method. However since
JS classes are finnicky, I had to mess around for a while to figure out a way
to define the logic in the base Model class, while providing the actual table
name in the derived class. This commit reflects my current efforts at doing
so in a syntactically and programmatically pleasing way, but there may be room for improvement.